Update BankAccount.java
From HW 6 update:
The account number should be a 9 digit number between 100 000 000 and 999 999 998 (no spaces). If a parameter is not a valid account number, set the value to 999 999 999 and print an error message to the screen about the problem. (Notice that 999,999,999 is not a valid value). This change should be reflected in your constructors as well as mutators.